+++ inc/manager.php
@@ -145,6 +145,9 @@ final class Optml_Manager {
 		if ( array_key_exists( 'et_fb', $_GET ) && ! empty( $_GET['et_fb'] ) ) {
 			return false; // @codeCoverageIgnore
 		}
+		if ( array_key_exists( 'tve', $_GET ) && $_GET['tve'] == 'true' ) {
+			return false; // @codeCoverageIgnore
+		}
 		if ( array_key_exists( 'context', $_GET ) && $_GET['context'] == 'edit' ) {
 			return false; // @codeCoverageIgnore
 		}
@@ -301,6 +304,7 @@ final class Optml_Manager {
 		$urls = array_map(
 			function ( $value ) {
 				$value = str_replace( '&quot;', '', $value );
+
 				return rtrim( $value, '\\";\'' );
 			},
 			$urls
@@ -464,7 +468,6 @@ final class Optml_Manager {
 		return array();
 	}
 
-
 	/**
 	 * Process url replacement from raw html strings.
 	 *
@@ -487,7 +490,7 @@ final class Optml_Manager {
 	 * @return array
 	 */
 	public function extract_image_urls_from_content( $content ) {
-		$regex = '/(?:http(?:s?):)(?:[\/\\\\|.|\w|\s|@|%|-])*\.(?:' . implode( '|', array_keys( Optml_Config::$extensions ) ) . ')(?:\?{1}[\w|=|&|\-|\.|:|;]*)?/';
+		$regex = '/(?:http(?:s?):)(?:[\/\\\\|.|\w|\s|-])*(?:[' . Optml_Config::$chars . '])*\.(?:' . implode( '|', array_keys( Optml_Config::$extensions ) ) . ')(?:\?{1}[\w|=|&|\-|\.|:|;]*)?/';
 		preg_match_all(
 			$regex,
 			$content,