+++ inc/url_replacer.php
@@ -116,20 +116,25 @@ final class Optml_Url_Replacer extends Optml_App_Replacer {
 
 		$is_slashed = strpos( $url, '\/' ) !== false;
 
-		$url = $is_slashed ? stripslashes( $url ) : $url;
+		// We do a little hack here, for json unicode chars we first replace them with html special chars,
+		// we then strip slashes to normalize the URL and last we convert html special chars back to get a clean URL
+		$url = $is_slashed ? html_entity_decode( stripslashes( preg_replace( '/\\\u([\da-fA-F]{4})/', '&#x\1;', $url ) ) ) : $url;
 
 		if ( strpos( $url, Optml_Config::$service_url ) !== false ) {
 			return $original_url;
 		}
+
 		if ( ! $this->can_replace_url( $url ) ) {
 			return $original_url;
 		}
+
 		// Remove any query strings that might affect conversion.
 		$url = strtok( $url, '?' );
 
 		if ( ! $this->is_valid_mimetype_from_url( $url, self::$filters[ Optml_Settings::FILTER_TYPE_OPTIMIZE ][ Optml_Settings::FILTER_EXT ] ) ) {
 			return $original_url;
 		}
+
 		if ( isset( $args['quality'] ) && ! empty( $args['quality'] ) ) {
 			$args['quality'] = $this->to_accepted_quality( $args['quality'] );
 		}
@@ -169,15 +174,21 @@ final class Optml_Url_Replacer extends Optml_App_Replacer {
 		}
 		$args = apply_filters( 'optml_image_args', $args, $original_url );
 
-		$new_url = ( new Optml_Image( $url, $args ) )->get_url(
-			[
-				'signed'          => $this->settings->use_lazyload() ? false : $this->is_allowed_site,
-				'apply_watermark' => apply_filters( 'optml_apply_watermark_for', true, $url ),
-			]
-		);
+		$arguments = [
+			'signed'          => $this->settings->use_lazyload() ? false : ! $this->is_allowed_site,
+			'apply_watermark' => apply_filters( 'optml_apply_watermark_for', true, $url ),
+
+		];
+
+		if ( isset( $args['format'] ) && ! empty( $args['format'] ) ) {
+			$arguments['format'] = $args['format'];
+		}
+
+		$new_url = ( new Optml_Image( $url, $args ) )->get_url( $arguments );
 
 		return $is_slashed ? addcslashes( $new_url, '/' ) : $new_url;
 	}
+
 
 	/**
 	 * Throw error on object clone